+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # 0.15.0
 
 - `rateLimit` - Added `maxItemsPerPeriod` option along with the options from `throughputLimiter`.
-- `throughputLimiter` - Added `minWindowLength`, `period`, and `expireAfter` options. The `windowLength` option is now `maxWindowLength`.
+- `throughputLimiter`
+    - Added `minWindowLength`, `period`, and `expireAfter` options.
+    - The `windowLength` option is now `maxWindowLength`.
+    - Separated `throttle` and `append`. These two convenience methods were also created: `throttleAndAppend` and `appendAndThrottle`.
 
 # 0.14.0