+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 # 0.15.0
 
 - `rateLimit` - Added `maxItemsPerPeriod` option along with the options from `throughputLimiter`.
-- `throughputLimiter`
-- Added `minWindowLength`, `period`, `expireAfter`, and `getTimeframe` options.
+- `throughputLimiter` - Added `minWindowLength`, `period`, `expireAfter`, and `getTimeframe` options.
     - The `windowLength` option is now `maxWindowLength`.
     - Separated `throttle` and `append`. These two convenience methods were also created: `throttleAndAppend` and `appendAndThrottle`.